+++ CHANGELOG.md
@@ -2,7 +2,7 @@ Changelog
 =========
 
 # Unreleased
-* Add `delay-close-send-stream` option which delays client send stream closure.
+* Add `stream-delay-close-send` option which delays client send stream closure.
 
 # 0.19.1 (2021-04-02)
 * Fix byte parsing to allow 8-bit signed integers to match the Thrift spec & other language implementations.